requirements.txt
python-dotenv==1.0.1
sqlalchemy==2.0.35
pandas==2.0.3
pyodbc==5.1.0
openpyxl==3.1.5